+++ src/utils/music/kw/lyric.js
@@ -1,19 +1,65 @@
 import { httpFetch } from '../../request'
+import { decodeName } from '../../index'
 
 export default {
   formatTime(time) {
-    const m = parseInt(time / 60)
-    const s = (time % 60).toFixed(2)
+    let m = parseInt(time / 60)
+    let s = (time % 60).toFixed(2)
     return (m < 10 ? '0' + m : m) + ':' + (s < 10 ? '0' + s : s)
   },
-  transformLrc({ songinfo, lrclist }) {
+  sortLrcArr(arr) {
+    const lrcSet = new Set()
+    let lrc = []
+    let lrcT = []
+
+    for (const item of arr) {
+      if (lrcSet.has(item.time)) {
+        const tItem = lrc.pop()
+        tItem.time = lrc[lrc.length - 1].time
+        lrcT.push(tItem)
+        lrc.push(item)
+      } else {
+        lrc.push(item)
+        lrcSet.add(item.time)
+      }
+    }
+
+    if (lrcT.length && lrc.length > lrcT.length) {
+      const tItem = lrc.pop()
+      tItem.time = lrc[lrc.length - 1].time
+      lrcT.push(tItem)
+    }
+
+    return {
+      lrc,
+      lrcT,
+    }
+  },
+  transformLrc(songinfo, lrclist) {
     return `[ti:${songinfo.songName}]\n[ar:${songinfo.artist}]\n[al:${songinfo.album}]\n[by:]\n[offset:0]\n${lrclist ? lrclist.map(l => `[${this.formatTime(l.time)}]${l.lineLyric}\n`).join('') : '暂无歌词'}`
   },
   getLyric(songId) {
     const requestObj = httpFetch(`http://m.kuwo.cn/newh5/singles/songinfoandlrc?musicId=${songId}`)
     requestObj.promise = requestObj.promise.then(({ body }) => {
-      if (body.status != 200) return Promise.reject(new Error('请求失败'))
-      return { lyric: this.transformLrc(body.data), tlyric: '', lxlyric: null }
+      // console.log(body)
+      if (!body.data?.lrclist?.length) return Promise.reject(new Error('Get lyric failed'))
+      let lrcInfo
+      try {
+        lrcInfo = this.sortLrcArr(body.data.lrclist)
+      } catch (err) {
+        console.log(err)
+        return Promise.reject(new Error('Get lyric failed'))
+      }
+      // console.log(body.data.lrclist)
+      // console.log(lrcInfo.lrc, lrcInfo.lrcT)
+      // console.log({
+      //   lyric: decodeName(this.transformLrc(body.data.songinfo, lrc)),
+      //   tlyric: decodeName(this.transformLrc(body.data.songinfo, lrcT)),
+      // })
+      return {
+        lyric: decodeName(this.transformLrc(body.data.songinfo, lrcInfo.lrc)),
+        tlyric: lrcInfo.lrcT.length ? decodeName(this.transformLrc(body.data.songinfo, lrcInfo.lrcT)) : '',
+      }
     })
     return requestObj
   },